refactor(expires_in): use util function
